There's a race condition at shutdown that this awkwardly works around...
it doesn't solve the generate problem of exiting the workflow when an app future is completed but task post-processing hasn't completed. see issue 1274

The test case added here sometimes records two entries in task fail history, and no end time; and
sometimes records three entries in task fail history and *does* have an end time.

In the first case, viz breaks because it cannot deal with tasks that have no end time.

This is a problem in writing data into the database: this should not behave differently database-wise between runs.

It was the last task in the run, in the run that didn't shut down right, which suggests maybe shutdown isn't working properly?